Move useSortRedirects to generic hooks dir
